moving matrix code around.  MutableMat is now in its own file, and that code only needs to be included by mat.cpp, soon to be mutablemat.cpp
git-svn-id: svn://svn.macaulay2.com/Macaulay2/branches/mike/jakob/jan2012-branch@14407 a8012f80-5f13-0410-8dc0-eb77bd940c0d

+++ Macaulay2/e/aring-ffpack.hpp
@@ -109,7 +109,7 @@ namespace M2 {
 
       void to_ring_elem(ring_elem &result, const ElementType &a) const
       {
-        result.int_val = a;
+        result.int_val = static_cast<int>(a);
       }
     
       void from_ring_elem(ElementType &result, const ring_elem &a) const
